RuleTest 수정
