+++ src/test/java/racingcar/model/CreateObjectCarsTest.java
@@ -41,4 +41,28 @@ public class CreateObjectCarsTest {
         assertEquals("CarB", carNamesArr[1]);
         assertEquals("CarC", carNamesArr[2]);
     }
+
+    @Test
+    @DisplayName("글자수가 5자 이하일 때 글자수 제한 오류 발생 여부")
+    void NonNameLenghtError() {
+        //Given
+        String carName = "carA";
+        CreateObjectCars createObjectCars = new CreateObjectCars();
+
+        //When & Then
+        assertDoesNotThrow(() -> createObjectCars.carNameLengthLimit(carName));
+    }
+
+    @Test
+    @DisplayName("글자수가 n자 초과일 때 글자수 제한 오류 발생 여부")
+    void nameLenghtError() {
+        //Given
+        String carName = "carABC";
+        CreateObjectCars createObjectCars = new CreateObjectCars();
+
+        //When & Then
+        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class, () -> createObjectCars.carNameLengthLimit(carName));
+        //Then
+        assertEquals("자동차 이름은 5자 이하만 가능합니다. 프로그램을 종료합니다.", exception.getMessage());
+    }
 }
